Do not error if header key/value is null
This is not an ideal change but if a header is a string and contains
a 0x0a (\n) within it then we end up with a string with no ":" this
change will truncate the header value, but will enable the payload
to continue processing.

This looks like an error in FUSE as the \r\n is not escaped.
    0x0260:  2042 756c 6b20 4f72 6465 7220 4974 656d  .Bulk.Order.Item
    0x0270:  2e0d 0a20 284f 5244 5f31 3031 2920 5b5d  ....(ORD_101).[]

TF-3231